id в БД сделал Int, т.к таким получаю его из сети

+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/data/entity/GenreEntity.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "genres")
 data class GenreEntity(
     @PrimaryKey
-    val gId: Long,
+    val gId: Int,
     val name: String
 )
